node-postgres integrated into back end, front end should be receiving expect data
