+++ client/src/App.jsx
@@ -10,7 +10,6 @@ import Guests from './components/Guests';
 
 import classes from './App.module.css';
 
-// do not GET and render calendar or price summary until clicked
 const CalendarCard = React.lazy(() => import('./components/CalendarCard'));
 const PriceSummary = React.lazy(() => import('./components/PriceSummary'));
 
@@ -21,11 +20,10 @@ export default function App() {
   const [checkIn, setCheckIn] = useState(null);
   const [checkOut, setCheckOut] = useState(null);
   const [home, setHome] = useState({
-    // Property data goes here:
     price: null,
     cleaning: null,
     avg: null,
-    totalRev: null,
+    totalrev: null,
   });
 
   useEffect(() => {
@@ -48,7 +46,6 @@ export default function App() {
     if (inOrOut === 'in') {
       setCheckIn(day);
     } else if (moment(checkIn, 'MM-DD-YYYY').diff(moment(day, 'MM-DD-YYYY')) < 0) {
-      // if checkIn date i
       setCheckOut(day);
     }
   };
@@ -77,7 +74,6 @@ export default function App() {
       <Suspense fallback={null}>
         <PriceSummary
           price={home.price}
-          // # of nights is difference between checkOut and checkIn date, in
           nights={moment(checkOut, 'MM-DD-YYYY').diff(moment(checkIn, 'MM-DD-YYYY'), 'days')}
           cleaning={home.cleaning}
         />
@@ -92,7 +88,7 @@ export default function App() {
     <div className={classes.container}>
       {showCalendar ? <div className={classes.background} onClick={toggleHandler} aria-hidden="true" /> : null}
       <div className={classes.price}>
-        <PriceReview price={home.price} avg={home.avg} total={home.totalRev} />
+        <PriceReview price={home.price} avg={home.avg} total={home.totalrev} />
       </div>
       <div className={classes.earliest}>
         <Earliest date="Apr 13" />
