+++ client/src/components/CalendarCard/index.jsx
@@ -24,7 +24,6 @@ export default function CalendarCard(props) {
   const [errMessage, setErrMessage] = useState('');
 
   useEffect(() => {
-    // GET calender ->
     axios.get('/calendar')
       .then((response) => {
         setCalendar(response.data);
