+++ server/controller/calendar.js
@@ -1,10 +1,9 @@
 const moment = require('moment');
-const { Booking } = require('../model');
+const db = require('../database');
 
-const cal = (req, res) => {
+const cal = (request, response) => {
   const calendarMonths = [];
   const dateObject = moment();
-  // generates calendar of up to two years from current date
   for (let i = 0; i < 24; i += 1) {
     const firstDayOfMonth = () => {
       const firstDay = moment(dateObject).startOf('month').format('d');
@@ -29,19 +28,12 @@ const cal = (req, res) => {
     dateObject.add(1, 'M');
   }
 
-  const { propertyId } = req.params;
-  Booking.find({ propertyId })
+  const { propertyId } = request.params;
+  db.query(`SELECT startDate, endDate FROM bookings WHERE propertyId = ${propertyId}`)
     .then((data) => {
-      // legacy: only 2 Bookings per Property in DB
-      // for each Booking, calculate the dates on the calendar that should are already booked
-      data.forEach(({ date }) => {
-        // moment expects start/end to be in "MM-DD-YYYY"
-        console.log(`date ${JSON.stringify(date)}`);
-
-        const dateStart = moment(date.start, 'MM-DD-YYYY');
-        const dateEnd = moment(date.end, 'MM-DD-YYYY');
-        console.log(`dateStart ${JSON.stringify(dateStart)}`);
-        console.log(`dateEnd ${JSON.stringify(dateEnd)}`);
+      data.rows.forEach((date) => {
+        const dateStart = moment(date.startdate, 'MM-DD-YYYY');
+        const dateEnd = moment(date.enddate, 'MM-DD-YYYY');
 
         const dayStart = +dateStart.format('D');
         const monthStart = dateStart.format('MMMM');
@@ -50,12 +42,10 @@ const cal = (req, res) => {
         const yearEnd = dateEnd.year();
 
         const daysDiff = dateEnd.diff(dateStart, 'days');
-        console.log(`daysDiff ${daysDiff}`);
         const firstDay = +dateStart.startOf('month').format('d');
         const firstEndDay = +dateEnd.startOf('month').format('d');
         const firstStartIndex = firstDay + dayStart - 1;
 
-        // check if booking goes thru end of one month and into next
         if (monthStart !== monthEnd) {
           const daysInFirstMonth = dateStart.daysInMonth() - (dayStart - 1);
           const daysInSecondMonth = daysDiff - daysInFirstMonth;
@@ -84,7 +74,7 @@ const cal = (req, res) => {
           });
         }
       });
-      res.send(calendarMonths);
+      response.send(calendarMonths);
     });
 };
