+++ server/controller/properties.js
@@ -4,7 +4,7 @@ const properties = (request, response) => {
   const { propertyId } = request.params;
   db.query(`SELECT * FROM properties WHERE propertyId = ${propertyId}`)
     .then(data => {
-      response.send(data)
+      response.send(data.rows[0]);
     });
 }
