+++ server/database/Dockerfile
@@ -1,3 +1,4 @@
+  
 FROM node:14-alpine
 
 WORKDIR /app
