+++ server/database/generateProperties.js
