--- server/model/index.js
@@ -1,4 +1,0 @@
-const Booking = require('./booking');
-const Property = require('./property');
-
-module.exports = { Booking, Property };
